Convert PIDs to be int(64) and count with a processor atomic

+++ runtime/include/chpl-privatization.h
@@ -8,8 +8,8 @@ extern void** chpl_privateObjects; // privatized array and domain objects
 
 extern void chpl_privatization_init(void);
 
-extern void chpl_newPrivatizedClass(void*, int32_t);
-extern void* chpl_getPrivatizedClass(int32_t);
+extern void chpl_newPrivatizedClass(void*, int64_t);
+extern void* chpl_getPrivatizedClass(int64_t);
 
 #endif // LAUNCHER
 #endif // _chpl_privatization_h_
